Initial Commit
Wanted to publish an old project. I wrote it for my Nokia N8. It uses a 4x4 grid, which is loaded from SQLite DB. Aim is to remove all pieces but one, capturing one piece each move.

Sample grid: "Q--------------Q"

There is a solver, also a generator (that does nothing). 

Don't know whether it is broken due to missing files.